some little changes added

+++ atelier/models/fabric.py
@@ -17,8 +17,7 @@ class Fabric(models.Model):
     ]
     name = models.CharField(max_length=264, verbose_name="Назва")
     group = models.CharField(max_length=3,choices=FABRIC_GROUPS, default=GROUP2, verbose_name="Група")
-    complexity_factor = models.DecimalField(max_digits=2, decimal_places=0, default=1.0,
-                                            verbose_name="Фактор складності")
+    complexity_factor = models.IntegerField( default=1, verbose_name="Фактор складності")
 
     def __str__(self):
         return self.name
